requirements.txt
torch
torchvision
torchaudio
pandas
rdkit
tqdm
selfies
torch_geometric
tensorboard
uv
pyyaml